# fix: potential deadlock in Subscriber