Fixes https://github.com/dunglas/mercure/pull/578/files#r748726493.