@divine 2021-11-13
Hello,

Does it affect performance in any case (again select case uh oh)? Benchmark results are the same, right?

I don't remember when it was added but I think subscribers were not closing correctly and memory was leaking or whatever...

Thanks!