@dunglas 2021-11-13
I didn't check performance yet because currently... It just doesn't work. The tests are still failing. I think that this case has been forgotten in the new implementation, it was one of the reasons my previous implementation was so complex. I hope that we'll manage to find a solution without partially reverting #578.